requirements.txt
absl-py==0.9.0
aiohttp @ file:///tmp/build/80754af9/aiohttp_1614360990960/work
alabaster==0.7.12
allennlp==0.9.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==20.3.0
audioread==2.1.8
Babel==2.7.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.3
blinker==1.4
blis==0.2.4
boto3==1.10.26
botocore==1.13.26
brotlipy==0.7.0
bs4==0.0.1
cachetools==4.0.0
certifi==2020.12.5
cffi==1.14.0
chardet==4.0.0
Click==7.0
colorama==0.4.1
conllu==1.3.1
cryptography @ file:///tmp/build/80754af9/cryptography_1613401388551/work
cycler==0.10.0
cymem==2.0.3
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
distro==1.5.0
docutils==0.15.2
editdistance==0.5.3
en-core-web-sm==2.1.0
entrypoints==0.3
ffmpeg-python==0.2.0
filelock==3.0.12
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
funcsigs==1.0.2
future==0.18.2
gast==0.3.2
gevent==1.4.0
google-auth==1.10.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.26.0
gym==0.10.9
h5py==2.10.0
-e git+https://github.com/facebookresearch/habitat-lab.git@4e7614f6b214aac99487bd9a172752891c7cd3ad#egg=habitat
habitat-sim==0.1.4
habitat==0.1.4
idna==2.10
idna-ssl @ file:///tmp/build/80754af9/idna_ssl_1611752490495/work
ifcfg==0.21
imageio==2.9.0
imageio-ffmpeg==0.4.3
imagesize==1.1.0
importlib-metadata==0.23
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonnet==0.14.0
jsonpickle==1.2
jsonschema==3.1.1
jupyter-client==6.1.11
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
librosa==0.8.0
llvmlite==0.31.0
magnum @ file:///vision/vision_users/sagnik/package_manager/habitat-sim/build/deps/magnum-bindings/src/python
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mir-eval==0.6
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==3.0.5
more-itertools==7.2.0
moviepy==1.0.3
multidict @ file:///tmp/build/80754af9/multidict_1607367768400/work
murmurhash==1.0.2
musdb==0.4.0
museval==0.4.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.1
nltk==3.4.5
notebook==6.2.0
numba==0.48.0
numpy==1.19.5
numpy-quaternion==2020.11.2.17.0.49
numpydoc==0.9.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
opt-einsum==3.1.0
overrides==2.5
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.3
parsimonious==0.8.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plac==0.9.6
pluggy==0.13.1
pooch==1.3.0
preshed==2.0.1
proglog==0.1.9
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.11.2
psutil==5.6.5
ptyprocess==0.7.0
py==1.8.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pycparser==2.20
pyglet==1.5.14
Pygments==2.4.2
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyroomacoustics==0.4.2
pyrsistent==0.15.5
PySocks @ file:///tmp/build/80754af9/pysocks_1605305763431/work
PySuperTuxKart==1.0a11
pytest==5.3.0
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.0
ray==0.7.6
redis==3.3.11
regex==2019.11.1
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
responses==0.10.6
rsa==4.0
s3transfer==0.2.1
scikit-build==0.11.1
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.5.0
sentencepiece==0.1.85
simplejson==3.17.2
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
spacy==2.1.9
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.2.0
stempeg==0.2.3
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
thinc==7.0.8
torch==1.4.0
torchaudio==0.4.0
torchsummary==1.5.1
torchvision==0.5.0
tornado==6.1
tqdm==4.56.0
traitlets==4.3.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
Unidecode==1.1.1
urllib3==1.26.3
wasabi==0.4.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
word2number==1.1
wrapt==1.12.1
yacs==0.1.8
yarl @ file:///tmp/build/80754af9/yarl_1606939915466/work
zipp==0.6.0
